backend/requirements.txt
crewai>=0.80.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
langchain-anthropic>=0.3.0
duckduckgo-search>=6.0.0
python-dotenv>=1.0.0
pymysql>=1.1.0
anthropic>=0.50.0
apscheduler>=3.10.0
feedparser>=6.0.0
